flask-mariadb/pyproject.toml
[tool.poetry]
name = "flask-mariadb"
version = "0.1.0"
description = ""
authors = ["chadcarlson <chad.carlson@platform.sh>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.1.2"
Flask-SQLAlchemy = "^2.5.1"
Flask-Migrate = "^3.1.0"
mysql-connector-python = "^8.0.29"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
